backend/pyproject.toml
[project]
name = "security-agent"
version = "0.1.0"
description = "Enterprise Security Incident Triage & Autonomous Runbook Agent - Kaggle AI Agents Capstone"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core Framework (minimal for API)
    "fastapi>=0.109.0,<0.120.0",
    "uvicorn[standard]>=0.27.0,<0.35.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    # HTTP & Async
    "httpx>=0.26.0,<0.29.0",
    "tenacity>=8.2.0,<10.0.0",
    # Utilities
    "python-dotenv>=1.0.0,<2.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
# LLM integration (Vertex AI / Gemini)
llm = [
    "langchain>=0.1.0,<0.4.0",
    "langchain-community>=0.0.10,<0.4.0",
    "langchain-google-vertexai>=1.0.0,<3.0.0",
    "langchain-core>=0.1.0,<0.4.0",
    "google-cloud-aiplatform>=1.40.0,<2.0.0",
    "google-adk",
    "google-genai",
]

# Database (PostgreSQL + Redis) - optional for production
db = [
    "asyncpg>=0.29.0,<0.31.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "pgvector>=0.2.4,<0.4.0",
    "redis>=5.0.0,<6.0.0",
    "upstash-redis>=1.0.0,<2.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
]

# Observability - optional
observability = [
    "structlog>=24.1.0,<25.0.0",
    "opentelemetry-api>=1.22.0,<2.0.0",
    "opentelemetry-sdk>=1.22.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0,<1.0.0",
    "langsmith>=0.0.87,<1.0.0",
]

# All features
all = [
    "security-agent[llm,db,observability]",
]

dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.25.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "api"]

[dependency-groups]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.25.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
